NB4444/pyproject.toml
[tool.poetry]
name = "nb4444"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.4.1"
backports-entry-points-selectable = "1.1.1"
beautifulsoup4 = "4.10.0"
certifi = "2021.10.8"
cffi = "1.15.0"
cfgv = "3.3.1"
chardet = "4.0.0"
coreapi = "2.3.3"
coreschema = "0.0.4"
cryptography = "35.0.0"
defusedxml = "0.7.1"
distlib = "0.3.3"
django = "~3.2"
django-autocomplete-light = "3.8.2"
django-bootstrap-v5 = "1.0.4"
django-cors-headers = "3.8.0"
django-templated-mail = "1.1.1"
django-webpack-loader = "1.4.1"
djangorestframework = "3.12.4"
djangorestframework-simplejwt = "4.8.0"
djoser = "2.1.0"
filelock = "3.4.0"
flake8 = "4.0.1"
gunicorn = "20.1.0"
identify = "2.4.0"
idna = "2.10"
importlib-metadata = "3.10.1"
install = "1.3.4"
itypes = "1.2.0"
jinja2 = "3.0.2"
markupsafe = "2.0.1"
mccabe = "0.6.1"
nodeenv = "1.6.0"
oauthlib = "3.1.1"
pillow = "9.0.1"
platformdirs = "2.4.0"
pre-commit = "2.15.0"
psycopg2-binary = "2.9.1"
pycodestyle = "2.8.0"
pycparser = "2.20"
pyflakes = "2.4.0"
pyjwt = "2.4.0"
python3-openid = "3.2.0"
pytz = "2020.1"
pyyaml = "6.0"
requests = "2.25.1"
requests-oauthlib = "1.3.0"
six = "1.16.0"
social-auth-app-django = "4.0.0"
social-auth-core = "4.1.0"
soupsieve = "2.2.1"
sqlparse = "0.3.1"
toml = "0.10.2"
uritemplate = "4.1.1"
urllib3 = "1.26.7"
virtualenv = "20.10.0"
whitenoise = "5.3.0"
zipp = "3.6.0"
django-el-pagination = "^3.3.0"
django-ckeditor = "^6.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
